Use longs instead of integers in elastic search

+++ app/lib/user_index.rb
@@ -69,11 +69,11 @@ class UserIndex < Chewy::Index
         end
         
         # https://www.elastic.co/guide/en/elasticsearch/reference/current/number.html
-        field :identity_id, type: "integer", include_in_all: false
+        field :identity_id, type: "long", include_in_all: false
         field :is_public, type: "boolean", include_in_all: false
         
         integer_columns.each do |column|
-          field(column.name.to_sym, type: "integer", include_in_all: false)
+          field(column.name.to_sym, type: "long", include_in_all: false)
         end
       end
